feat: reorganize project structure
- Move python_baseline/ to src/ with clean structure
- Separate API routers (baseline, mysolver)
- Move models to domain/
- Update all imports from python_baseline.* to src.*
- Remove legacy files and duplicates
- Improve project organization and maintainability
