--- python_baseline/README.md
@@ -1,15 +1,0 @@
-# Python Baseline Solver
-
-This directory contains the baseline implementation of the CoW Protocol solver, ported from Rust to Python.
-
-## Directory Structure
-
-```
-python_baseline/
-├── api/           # HTTP API endpoints
-├── models/        # Domain models (Auction, Order, Solution, etc.)
-├── utils/         # Utility functions (math, serialization, etc.)
-├── infra/         # Infrastructure (settings, metrics, logging, CLI)
-├── tests/         # Test cases
-├── README.md      # This file
-```
