--- python_baseline/__init__.py
@@ -1,9 +1,0 @@
-"""
-CoW Protocol Baseline Solver
-
-This package contains the baseline implementation of the CoW Protocol solver,
-ported from Rust to Python.
-"""
-
-__version__ = "1.0.0"
-__author__ = "CoW Protocol Team"
