--- python_baseline/api/__init__.py
@@ -1,3 +1,0 @@
-"""
-API layer for the baseline solver.
-"""
