--- python_baseline/engines/mysolver/engine.py
@@ -1,68 +1,0 @@
-"""
-MySolver engine implementation.
-"""
-
-import logging
-from python_baseline.models.auction import Auction
-from python_baseline.models.solution import Solution, Solutions
-from python_baseline.engines.base import SolverEngine
-
-
-class MySolverEngine:
-    """
-    MySolver engine implementation.
-    
-    This is a stub implementation for user-defined solver logic.
-    Returns minimal valid solutions.
-    """
-    
-    def __init__(self):
-        """Initialize the MySolver engine."""
-        self.logger = logging.getLogger(f"{__name__}.MySolverEngine")
-    
-    async def solve(self, auction: Auction) -> Solutions:
-        """
-        Solve the auction and return solutions.
-        
-        Args:
-            auction: The auction to solve
-            
-        Returns:
-            Solutions object containing the solver's solutions
-        """
-        self.logger.info(f"Solving auction {auction.id} with MySolver engine")
-        self.logger.info(f"Orders: {len(auction.orders)}, Tokens: {len(auction.tokens)}")
-        
-        # Stub implementation - return minimal valid solution
-        solution = Solution(
-            id=auction.id,
-            trades=[],
-            prices={},
-            interactions=[]
-        )
-        
-        return Solutions(solutions=[solution])
-    
-    def _validate_auction(self, auction: Auction) -> bool:
-        """
-        Validate the auction data.
-        
-        Args:
-            auction: The auction to validate
-            
-        Returns:
-            True if valid, False otherwise
-        """
-        # Basic validation
-        if not auction.id:
-            return False
-        
-        if not auction.orders:
-            self.logger.warning("Auction has no orders")
-            return False
-        
-        if not auction.tokens:
-            self.logger.warning("Auction has no tokens")
-            return False
-        
-        return True
