+++ src/api/routers/__init__.py
@@ -1,0 +1,3 @@
+"""
+API routers package.
+"""
