+++ src/domain/common.py
