+++ src/domain/solution.py
