+++ src/infra/di.py
@@ -2,9 +2,9 @@
 Dependency injection for solver engines.
 """
 
-from python_baseline.engines.baseline.engine import BaselineEngine
-from python_baseline.engines.mysolver.engine import MySolverEngine
-from python_baseline.engines.base import SolverEngine
+from src.engines.baseline.engine import BaselineEngine
+from src.engines.mysolver.engine import MySolverEngine
+from src.engines.base import SolverEngine
 
 
 def baseline_engine() -> SolverEngine:
