--- src/models/legacy/batch_auction.py
@@ -1,43 +1,0 @@
-"""
-LEGACY: Batch Auction model for reference.
-This file is kept for reference and shows the old structure.
-DO NOT USE in new implementations - use cow_auction.py instead.
-"""
-from __future__ import annotations
-
-import decimal
-import logging
-from decimal import Decimal
-from typing import Any, Optional
-
-# This is the OLD structure - kept for reference only
-# Use cow_auction.py for new implementations
-
-class BatchAuction:
-    """
-    LEGACY: Model containing BatchAuction which is what solvers operate on.
-    
-    This is kept for reference only. The new structure uses:
-    - cow_auction.py for auction data
-    - cow_solution.py for solution data
-    """
-    
-    def __init__(self, name: str = "batch_auction"):
-        self.name = name
-        self.orders = []
-        self.tokens = {}
-        self.liquidity = []
-        
-    @classmethod
-    def from_dict(cls, data: dict, name: str) -> BatchAuction:
-        """Create BatchAuction from dictionary - LEGACY METHOD"""
-        logging.warning("Using legacy BatchAuction.from_dict - consider migrating to CowAuction")
-        batch = cls(name)
-        # Legacy implementation would go here
-        return batch
-        
-    def solve(self):
-        """Solve the batch auction - LEGACY METHOD"""
-        logging.warning("Using legacy BatchAuction.solve - implement in /solve endpoint instead")
-        # Legacy implementation would go here
-        pass
