+++ src/tests/test_api_solve_contract.py
@@ -6,7 +6,7 @@ This module contains tests to verify the /solve endpoint contract compliance.
 
 import pytest
 from fastapi.testclient import TestClient
-from python_baseline.api.app import app
+from src.api.app import app
 
 
 client = TestClient(app)
